Fix incorrect cast in ts_language_symbol_is_in_progress

+++ include/tree_sitter/parser.h
@@ -83,7 +83,7 @@ typedef union {
 
 typedef union {
   TSSymbol symbol;
-  unsigned int count;
+  unsigned short count;
 } TSInProgressSymbolEntry;
 
 struct TSLanguage {
